Add fritzing file
This is the first version, pushbutton and optocoupler for each focus and
shutter are next to each other, in the middle there's the resistor
triarchy that controls the remote port. Focus is controlled by the
button/optocoupler to the right, Shutter by the button/optocoupler to
the left. The button to the far right is used to start/top the
intervalometer.