Change to easier layout and configurable interval
The resistor triarchy is not so central anymore, and pushbuttons and
optocouplers aren't grouped by function but by type now.

The wiring to the switches is neater now since they use the ground row
at the top of the breadboard.

Pins 0-7 are now input pins to control the interval speed. They
represent a binary system, each pin that's connected to the ground row
at the bottom of the breadboard is counted as on and influences the
speed.

The button to the far left starts/stops the intervalometer. The LED
indicates this state.